refactor: merge 2 from dev
